+++ src/IntervalAlgebra/IntervalUtilities.hs
@@ -17,12 +17,18 @@ module IntervalAlgebra.IntervalUtilities (
     , clip
     , relations
     , gapsWithin
+    , emptyIf
+    , emptyIfNone
+    , emptyIfAny
+    , emptyIfAll
 ) where
 
 import GHC.Base
-    ( (++), map, foldr, otherwise, ($), (.), (<*>), seq
-    , Semigroup((<>)), Functor(fmap), Maybe(..), Int)
-import Data.Foldable()
+    ( (++), map, foldr, otherwise, ($), (.), (<*>), seq, not
+    , Semigroup((<>)), Functor(fmap), Maybe(..)
+    , Int, Bool)
+import Data.Foldable ( Foldable(null, foldl'), all, any )
+import Data.Monoid ( (<>), Monoid(mempty) )
 import Prelude (uncurry, zip, Num, foldl)
 import IntervalAlgebra
     ( Interval, Intervallic(..), IntervalAlgebraic(..)
@@ -30,13 +36,14 @@ import IntervalAlgebra
     , IntervalFilterable(..)
     , IntervalRelation(..))
 import Data.Maybe (mapMaybe)
-import Data.List ( (++), null, any, head, init, last, tail )
+import Data.List ( (++), head, init, last, tail )
 import Witherable ( Filterable )
 
 intInt :: Int -> Int -> Interval Int
 intInt = unsafeInterval
 
 -- | Box to avoid overlapping instances
+-- TODO: avoid the head/tail footguns
 newtype Box a = Box { unBox :: [a] }
 instance (IntervalCombinable a) => Semigroup (Box (Interval a)) where
     Box x <> Box y
@@ -55,10 +62,10 @@ instance (IntervalCombinable a) => Semigroup (Box (Interval a)) where
 combineIntervals :: (IntervalCombinable a) => [Interval a] -> [Interval a]
 combineIntervals l = unBox $ foldl' (<>) (Box []) (map (\z -> Box [z]) l)
 
-foldl' :: (t -> a -> t) -> t -> [a] -> t
-foldl' f z []     = z
-foldl' f z (x:xs) = let z' = z `f` x 
-                     in seq z' $ foldl' f z' xs
+-- foldl' :: (t -> a -> t) -> t -> [a] -> t
+-- foldl' f z []     = z
+-- foldl' f z (x:xs) = let z' = z `f` x
+--                      in seq z' $ foldl' f z' xs
 
 -- | Returns a (possibly empty) list of intervals consisting of the gaps between
 --   intervals in the input list. *To work properly, the input list should be sorted*.
@@ -110,6 +117,49 @@ gapsWithin :: (IntervalSizeable a b, IntervalCombinable a, IntervalFilterable []
   -> [Interval a] -- ^ x
   -> [Interval a]
 -- TODO: generalize to collections besides list
-gapsWithin i x = gaps $ enderval 0 (begin i) : 
+gapsWithin i x = gaps $ enderval 0 (begin i) :
                         mapMaybe (clip i) (filterNotDisjoint i x) ++
-                        [beginerval 0 (end i)] 
+                        [beginerval 0 (end i)]
+
+-- | Given a predicate combinator, a predicate, and list of intervals, returns 
+--   the input unchanged if the predicate combinator is 'True'. Otherwise, returns
+--   an empty list. See 'emptyIfAny' and 'emptyIfNone' for examples.
+emptyIf :: (Monoid (f (Interval a)), Foldable f, IntervalFilterable f a)=>
+     ((Interval a -> Bool) -> f (Interval a) -> Bool) -- ^ e.g. 'any' or 'all'
+  -> (Interval a -> Bool) -- ^ predicate to apply to each element of input list
+  -> f (Interval a)
+  -> f (Interval a)
+emptyIf g f x = if g f x then mempty else x
+
+-- | Returns the empty monoid structure if *none* of the element of input satisfy
+--   the predicate condition.
+-- 
+-- For example, the following returns the empty list because none of the intervals
+-- in the input list 'starts' (3, 5).
+-- >>> emptyIfNone (starts (intInt 3 5)) [intInt 3 4, intInt 5 6]
+-- []
+--
+-- In the following, (3, 5) 'starts' (3, 6), so the input is returned.
+-- >>> emptyIfNone (starts (intInt 3 5)) [intInt 3 6, intInt 5 6]
+-- [(3, 6),(5, 6)]
+emptyIfNone :: (Monoid (f (Interval a)), Foldable f, IntervalFilterable f a)=>
+    (Interval a -> Bool) -- ^ predicate to apply to each element of input list
+  -> f (Interval a)
+  -> f (Interval a)
+emptyIfNone = emptyIf (\f x -> (not.any f) x)
+
+-- | Returns the empty monoid structure if *any* of the element of input satisfy
+--   the predicate condition
+emptyIfAny :: (Monoid (f (Interval a)), Foldable f, IntervalFilterable f a)=>
+    (Interval a -> Bool) -- ^ predicate to apply to each element of input list
+  -> f (Interval a)
+  -> f (Interval a)
+emptyIfAny = emptyIf any
+
+-- | Returns the empty monoid structure if *all* of the element of input satisfy
+--   the predicate condition
+emptyIfAll :: (Monoid (f (Interval a)), Foldable f, IntervalFilterable f a)=>
+    (Interval a -> Bool) -- ^ predicate to apply to each element of input list
+  -> f (Interval a)
+  -> f (Interval a)
+emptyIfAll = emptyIf all